Split settings.py
Also move functions specific to "region math" into a new sub-module and
add `get_settings` function to lib for all other plugins to fetch from.
